+++ Example-iOS/ViewControllers/ExampleViewController.swift
@@ -29,6 +29,18 @@ import QuickTableViewController
 
 internal final class ExampleViewController: QuickTableViewController {
 
+  init() {
+    if #available(iOS 13.0, *) {
+      super.init(style: .insetGrouped)
+    } else {
+      super.init(style: .grouped)
+    }
+  }
+
+  required init?(coder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+
   // MARK: - Properties
 
   private let debugging = Section(title: nil, rows: [NavigationRow(text: "", detailText: .none)])