+++ Example-iOS/ViewControllers/RootViewController.swift
@@ -29,14 +29,28 @@ import QuickTableViewController
 
 internal final class RootViewController: QuickTableViewController {
 
-  override func viewDidLoad() {
-    super.viewDidLoad()
+  init() {
+    if #available(iOS 13.0, *) {
+      super.init(style: .insetGrouped)
+    } else {
+      super.init(style: .grouped)
+    }
 
     let titleLabel = UILabel()
     titleLabel.text = "QuickTableViewController"
     titleLabel.font = UIFont.boldSystemFont(ofSize: 17)
     title = " "
     navigationItem.titleView = titleLabel
+  }
+
+  required init?(coder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+
+  // MARK: - UIViewController
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
 
     tableContents = [
       Section(title: "Default", rows: [